phase1-ios/ModelConversion/requirements.txt
pillow
torch
torchvision
transformers
coremltools
kornia
timm
